feat: Add rhel81-power8 support
Combined the static host tests into a single scenario, which tests to
make sure it downloads the node_exporter and checks to make sure the
service runs on the host by curling the /metrics endpoint.

To test locally, make sure to add any hosts you want into your
~/.ssh/config.

This systemd set up should work on any system that works with systemd.
It runs node_exporter as the root user and saves the service file into
/etc/systemd/system and the config file is saved to
/etc/sysconfig/node_exporter.

The tests will run on port 9101, so that it does not conflict with any
existing instance of node_exporter running on the default port 9100.
Also it saves the systemd file as node_exporter_test so that it does
not conflict with any exising node_exporter systemd service.